requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
sentence-transformers==3.2.1
groq==0.11.0
llama-index==0.11.18
llama-index-vector-stores-supabase==0.2.2
supabase==2.5.0
pydantic==2.9.2
pydantic-settings==2.11.0
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.1.0
